Restructured ui related components to ui package
